+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Instrumentation names are changed (possibly breaking for benchmarks records)
 - Temporary performance loss is expected in orded to keep compatibility between `Variable` and `Parameter` paradigms.
 - `FolderFunction` must now be accessed through `nevergrad.parametrization.FolderFunction`
+- `PSO` now uses initialization by sampling the parametrization, instead of sampling all the real space.
 
 ## v0.3.0